requirements.txt
openai>=1.0.0
clickhouse-connect>=0.6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0


